requirements.txt
flask
flask-cors
gunicorn
tensorflow
numpy==1.26.4
pillow
opencv-python
mysql-connector-python
